Production: Netlify Functions deployment
- Clean Netlify Functions version
- Serverless backend with Gemini AI
- Optimized for Netlify deployment
- Ready for production deployment
