+++ README.md
@@ -1,77 +1,72 @@
 # Joke Generator
 
-A minimal, clean web application that generates random jokes based on user input. Built with HTML, CSS, and vanilla JavaScript following production-grade coding practices.
+AI-powered joke generator using Gemini API with Netlify Functions deployment.
 
 ## Features
 
-- 🎭 Clean, modern UI with gradient backgrounds and smooth animations
-- 📱 Fully responsive design that works on all devices
-- ✅ Input validation with real-time feedback
-- ⚡ Loading states and error handling
-- 🎨 Beautiful animations and transitions
-- 🧹 Clean, well-documented code
+- 🎭 AI-generated jokes using Gemini API
+- 📊 Token usage tracking
+- ⚡ Serverless backend with Netlify Functions
+- 📱 Responsive design
+- 🧹 Clean, production-ready code
 
-## How to Run
+## How to Deploy
 
-1. Simply open `index.html` in your web browser
-2. Enter a word or sentence in the input field
-3. Click "Generate Joke" to get a random joke!
+1. **Connect to Netlify:**
+   - Go to [Netlify.com](https://netlify.com)
+   - Connect your GitHub repository
+   - Select the `production` branch
+
+2. **Add Environment Variables:**
+   - `GEMINI_API_KEY` = your Gemini API key
+
+3. **Deploy:**
+   - Netlify will automatically deploy your app
+   - Your app will be available at `https://your-app-name.netlify.app`
 
 ## Project Structure
 
 ```
-├── index.html      # Main HTML structure
-├── styles.css      # CSS styling and animations
-├── script.js       # JavaScript functionality
-└── README.md       # This file
+├── index.html              # Frontend
+├── styles.css              # Frontend styling
+├── script.js               # Frontend JavaScript
+├── netlify/
+│   └── functions/
+│       └── joke.js         # Netlify Function
+├── netlify.toml            # Netlify configuration
+├── package.json            # Dependencies
+└── README.md               # This file
 ```
 
-## Code Quality Features
+## Local Development
 
-### HTML
-- Semantic HTML5 structure
-- Proper accessibility attributes
-- Clean, readable markup
+1. **Install dependencies:**
+   ```bash
+   npm install
+   ```
 
-### CSS
-- Modern CSS with flexbox and grid
-- Responsive design with mobile-first approach
-- CSS custom properties for maintainability
-- Smooth animations and transitions
-- Clean, organized stylesheet structure
+2. **Open in browser:**
+   ```bash
+   # Open index.html in your browser
+   # The app will use Netlify Functions when deployed
+   ```
 
-### JavaScript
-- ES6+ features and modern syntax
-- Comprehensive error handling
-- Input validation
-- Clean separation of concerns
-- Well-documented functions
-- Production-ready code structure
-- Async/await for better readability
+## Environment Variables
 
-## Learning Points
+Create `.env` file for local development:
+```env
+GEMINI_API_KEY=your_gemini_api_key_here
+```
 
-This project demonstrates:
+## Technologies Used
 
-1. **Clean Code Principles**: Well-organized, readable code with clear function names
-2. **Error Handling**: Proper validation and user feedback
-3. **Modern JavaScript**: ES6+ features, async/await, proper event handling
-4. **Responsive Design**: Mobile-first CSS approach
-5. **User Experience**: Loading states, animations, and clear feedback
-6. **Code Organization**: Separation of concerns and modular structure
+- **Frontend:** HTML, CSS, JavaScript
+- **Backend:** Netlify Functions (serverless)
+- **AI:** Google Gemini API
+- **Deployment:** Netlify
 
-## Browser Compatibility
+## Cost
 
-Works in all modern browsers that support:
-- ES6+ JavaScript features
-- CSS Grid and Flexbox
-- CSS Custom Properties
-- Async/await
-
-## Future Enhancements
-
-- Connect to a real joke API
-- Add more joke categories
-- Implement joke rating system
-- Add sharing functionality
-- Include more animations
+- **Netlify Functions:** Free tier (125k calls/month)
+- **Gemini API:** Free tier available
+- **Total:** $0/month for personal use
