+++ package.json
@@ -1,22 +1,16 @@
 {
-  "name": "joke-generator-backend",
+  "name": "joke-generator",
   "version": "1.0.0",
-  "description": "Backend API for joke generator with Gemini integration",
-  "main": "server.js",
+  "description": "AI-powered joke generator with Netlify Functions",
+  "main": "index.html",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "build": "echo 'No build step needed for static site'",
+    "dev": "echo 'Open index.html in browser for development'"
   },
   "dependencies": {
-    "express": "^4.18.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
     "@google/generative-ai": "^0.2.1"
   },
-  "devDependencies": {
-    "nodemon": "^3.0.1"
-  },
-  "keywords": ["joke", "api", "gemini", "express"],
+  "keywords": ["joke", "ai", "gemini", "netlify", "serverless"],
   "author": "Your Name",
   "license": "MIT"
 }
